+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Remove unnecessary "Sell price (Flick)" info widget on sea creature details screen, as Flick does not purchase sea creatures.
 
 ## [0.7.2] - 2020-11-11
 ### Added