+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [0.8.0] - 2020-11-22
 ### Added
 - Add new badges/chips to show if an item is newly available (was not available previous month) or leaving soon (will not be available next month).
 - Add hero transition when navigating between home screen and details screen. Thumbnail/image widget now transitions between positions on those screens.